added tests for Esprima issue
Note! I didn't fix on purpose because NodeJS does it automatically.

However. It's an easy task 1) pass the same arg to all functions in "pattern.ts" or 2) store an obj / array on the parser object and take it from there (*then no need to parse as args*)

This is related to scoping - somethig I'm not implementing now.