reduced code size and optimized JSX
I optimized JSX, and  was playing around with it after reading the readme on the Acorn JSX plugin repo.
I run several benchmaks and discovered that something has to be horrible wrong with that plugin .

In microbenchmark it's average is 27k opts / sec for simple cases like `<a>></a>`.

Esprima's average is 300k opts/sec.

Cherow's average is around 1 million opts/sec

- J.K Thomas (drop-by-dev)