refactor(cherow): simplify lexer code (#202)
* refactor(cherow) fixes multiLine comment spec deviation

`/* */ -->` - equal to multi-line comment without a newline - shouldn't be treated  as an HTML comment, but throw in the parser. Both Acorn and Esprima suffers from this.

There exist an 'onToken' test for this.

* refactor(cherow): improve hot path number scanning

* refactor(cherow): added back some tests

* refactor(cherow): fixing raw property

From v. 2.0 the raw property will be on the AST node by default, but set as null. if `raw` option is set, it will get an value

* refactor(cherow): fixing broken lookahead tests

* refactor(cherow) fix BigInt not failing on MV as a none-integer value

* refactor(cherow): simplified number scanning