refactor(cherow): fixing correct error locations (#205)
* refactor(cherow): fixing correct error locations

Example `/* asf  `. In Acorn the error loc is reported as `1:0`. In Esprima `Line 1`
and in ShiftJS reported as  `1:8`.  None of this is correct if you try to count.  There are nothing on pos 8 as reported by ShiftJS. The last char is a WS on pos 7.

in this case Cherow reports the error location like this `1:7`.  Counted from comment start to the last WS.

Another example is this one `"foo `. Wrong in all parsers, but ShiftJS got it almost right again. They are reporting `1:5` as the current pos. Note:  The last pos is 4, and pos 5 as reported is non-existing.

In this case Cherow reports this as `1:3`. The reason it stops at pos 3, is because the `o` should have been the closing quote. Note: This is an unterminated string. This can be adjusted if my thinking is wrong, but at least more correct than other parsers

* refactor(cherow): added more test and fixed options

* refactor(cherow): fix style issues

* refactor(cherow): removed duplicate tests