refactor(cherow): enabled block scoping
Here I added the basic for block scoping. As seen in the code every time the 'parseBlockStatement' are invoked we are going one level down and create nested scope.  For other part of the code we need to create a "sub scope" that re-use the parent scope. There are no big performance concerns involved in this.